requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.5.0
pymongo==4.6.1
pydantic==2.8.0
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
openai==1.12.0
email-validator==2.1.0.post1
python-dotenv==1.0.1
aiofiles==23.2.1
twilio==8.1.0
sendgrid==6.11.0